修正findMin()和findMinMax()函数的位置返回类型
size_t是unsigned 类型的，把int赋值给一个unsigned类型不合法，将会导致错误

+++ chapter9/9-1/findMinMAxNumber/findMinMax.cc
@@ -32,8 +32,8 @@ class MinMaxPair
 {
 public:
     MinMaxPair(int _minPos = -1 , int _maxPos = -1):minPos(_minPos),maxPos(_maxPos){}
-    size_t minPos;
-    size_t maxPos;
+    int minPos;
+    int maxPos;
     bool operator==(const MinMaxPair & pair)
     {
         return (this->minPos == pair.minPos && this->maxPos == pair.maxPos);
